Player
added the player, it jumps like a pogostick. the sprite changes. it uses the new input system in unity (that changes nothing if you still just set an action to a key). added layer called "ground" which is used to check if the player is standing on ground and can thus jump